utils/jwt_utils.py
import jwt
from datetime import datetime, timedelta

SECRET_KEY = "super-secret-key"  # CHANGE IN PRODUCTION

def create_token(user_id):
    payload = {
        "user_id": user_id,
        "exp": datetime.utcnow() + timedelta(days=7)
    }
    return jwt.encode(payload, SECRET_KEY, algorithm="HS256")

def decode_token(token):
    try:
        return jwt.decode(token, SECRET_KEY, algorithms=["HS256"])
    except:
        return None
