requirements.txt
flask
flask-cors
python-dotenv
gunicorn
PyJWT
werkzeug
flask-jwt-extended
